[0.5.0-SNAPSHOT] Injection fixed Tests cleanup
