--- cassandra/src/test/java/io/goodforgod/testcontainers/extensions/cassandra/ContainerPerClassTests.java
@@ -1,36 +1,0 @@
-package io.goodforgod.testcontainers.extensions.cassandra;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import io.goodforgod.testcontainers.extensions.ContainerMode;
-import org.junit.jupiter.api.*;
-
-@TestcontainersCassandra(mode = ContainerMode.PER_CLASS, image = "cassandra:4.1")
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class ContainerPerClassTests {
-
-    @ContainerCassandraConnection
-    private CassandraConnection sameConnection;
-
-    private static CassandraConnection firstConnection;
-
-    @Order(1)
-    @Test
-    void firstConnection(@ContainerCassandraConnection CassandraConnection connection) {
-        assertNull(firstConnection);
-        assertNotNull(connection);
-        assertNotNull(sameConnection);
-        assertEquals(sameConnection, connection);
-        firstConnection = connection;
-    }
-
-    @Order(2)
-    @Test
-    void secondConnection(@ContainerCassandraConnection CassandraConnection connection) {
-        assertNotNull(connection);
-        assertNotNull(firstConnection);
-        assertNotNull(sameConnection);
-        assertEquals(sameConnection, connection);
-        assertEquals(firstConnection, connection);
-    }
-}
