--- mariadb/src/test/java/io/goodforgod/testcontainers/extensions/mariadb/MariadbContainerPerClassTests.java
@@ -1,43 +1,0 @@
-package io.goodforgod.testcontainers.extensions.mariadb;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import io.goodforgod.testcontainers.extensions.ContainerMode;
-import io.goodforgod.testcontainers.extensions.jdbc.ContainerMariadbConnection;
-import io.goodforgod.testcontainers.extensions.jdbc.JdbcConnection;
-import io.goodforgod.testcontainers.extensions.jdbc.TestcontainersMariadb;
-import org.junit.jupiter.api.*;
-
-@TestcontainersMariadb(mode = ContainerMode.PER_CLASS, image = "mariadb:11.0-jammy")
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class MariadbContainerPerClassTests {
-
-    @ContainerMariadbConnection
-    private JdbcConnection sameConnection;
-
-    private static JdbcConnection firstConnection;
-
-    @Order(1)
-    @Test
-    void firstConnection(@ContainerMariadbConnection JdbcConnection connection) {
-        assertNull(firstConnection);
-        assertNotNull(connection);
-        assertNotNull(connection.params().jdbcUrl());
-        assertNotNull(sameConnection);
-        assertEquals(sameConnection, connection);
-        firstConnection = connection;
-        connection.execute("SELECT 1;");
-    }
-
-    @Order(2)
-    @Test
-    void secondConnection(@ContainerMariadbConnection JdbcConnection connection) {
-        assertNotNull(connection);
-        assertNotNull(connection.params().jdbcUrl());
-        assertNotNull(firstConnection);
-        assertNotNull(sameConnection);
-        assertEquals(sameConnection, connection);
-        assertEquals(firstConnection, connection);
-        connection.execute("SELECT 1;");
-    }
-}
