--- redis/src/test/java/io/goodforgod/testcontainers/extensions/redis/ContainerPerClassTests.java
@@ -1,36 +1,0 @@
-package io.goodforgod.testcontainers.extensions.redis;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import io.goodforgod.testcontainers.extensions.ContainerMode;
-import org.junit.jupiter.api.*;
-
-@TestcontainersRedis(mode = ContainerMode.PER_CLASS, image = "redis:7.0-alpine")
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class ContainerPerClassTests {
-
-    @ContainerRedisConnection
-    private RedisConnection sameConnection;
-
-    private static RedisConnection firstConnection;
-
-    @Order(1)
-    @Test
-    void firstConnection(@ContainerRedisConnection RedisConnection connection) {
-        assertNull(firstConnection);
-        assertNotNull(connection);
-        assertNotNull(sameConnection);
-        assertEquals(sameConnection, connection);
-        firstConnection = connection;
-    }
-
-    @Order(2)
-    @Test
-    void secondConnection(@ContainerRedisConnection RedisConnection connection) {
-        assertNotNull(connection);
-        assertNotNull(firstConnection);
-        assertNotNull(sameConnection);
-        assertEquals(sameConnection, connection);
-        assertEquals(firstConnection, connection);
-    }
-}
